+++ src/index.ts
@@ -1,5 +1,5 @@
 export * from './colors';
 export * from './css';
+export * from './makeVariables';
 export * from './scoped';
 export * from './units';
-export * from './variables';